+++ src/main/kotlin/kr/hhplus/be/server/common/exception/Exceptions.kt
@@ -3,6 +3,8 @@ package kr.hhplus.be.server.common.exception
 // Coupon
 class InvalidCouponStatusException: BusinessException(ErrorCode.INVALID_COUPON_STATUS)
 class ExpiredCouponException: BusinessException(ErrorCode.EXPIRED_COUPON)
+class ExceededMaxCouponLimitException: BusinessException(ErrorCode.EXCEEDED_MAX_COUPON_LIMIT)
+
 
 // Payment
 class AlreadyPaidException: BusinessException(ErrorCode.ALREADY_PAID)
