+++ src/main/kotlin/kr/hhplus/be/server/coupon/domain/model/Coupon.kt
@@ -1,7 +1,9 @@
 package kr.hhplus.be.server.coupon.domain.model
 
 import jakarta.persistence.*
-import kr.hhplus.be.server.coupon.domain.ExceededMaxCouponLimitException
+import kr.hhplus.be.server.common.exception.ExceededMaxCouponLimitException
+import kr.hhplus.be.server.order.domain.Order
+import kr.hhplus.be.server.order.domain.OrderItem
 import java.math.BigDecimal
 import java.time.LocalDateTime
 
@@ -70,6 +72,19 @@ class Coupon(
 
     fun isApplicableTo(context: DiscountContext): Boolean {
         return this.discountPolicy.discountCondition.isSatisfiedBy(context)
+    }
+
+    fun applicableItems(order: Order, userId: Long): List<OrderItem> {
+        return order.orderItems.filter {
+            isApplicableTo(
+                DiscountContext(
+                    userId = userId,
+                    productId = it.productId,
+                    variantId = it.variantId,
+                    orderAmount = order.originalTotal
+                )
+            )
+        }
     }
 
     fun issueTo(userId: Long, now: LocalDateTime = LocalDateTime.now()): UserCoupon {
